requirements.txt
python-telegram-bot==22.0
httpx~=0.27
apscheduler==3.11.0
pytz==2025.1
tzlocal==5.3.1
pymongo
requests
